Implemented MarketProtocol, modified UserManagaerProcotol and UserManager
